code update

+++ lib/pages/students/views/add_new_lesson/type_card.dart
@@ -12,7 +12,7 @@ class ManageStudentLessonTypeCard extends StatelessWidget {
   });
 
   final int lessonCount;
-  final int hoursCount;
+  final double hoursCount;
   final int type;
   final bool selected;
   final Function ontap;
